Change how fedora user credentials are handled
This commit removes the fcrepo username, which wasn't used by the template and also the direct input of Access and secret keys. The keys now live in a yaml file and there is a new variable called fcrepo_binary_bucket_yaml_file which is required.
This variable is a path to the file containing the aws fcrepo access and secret keys. Check the README for details.